+++ docs/source/changelog.md
@@ -3,7 +3,7 @@
 ## 0.9.10 (2025-mm-dd)
 - Fix incompatibility with pydiverse.transform >=0.2.1 (<0.2.0 still supported) (0.2.0 will not be supported):
   - pydiverse transform had a radical refactoring around this point
-- NotImplementedError("pydiverse.transform 0.2.0 isn't supported") when pydiverse.transform 0.2.0 is being used.
+- NotImplementedError when pydiverse.transform 0.2.0 - 0.2.2 is being used.
 - Pinned prefect to version ">=2.13.5, <3.0.0", because future versions are currently not supported
 - IBM DB2 tests: switch from pypi to conda-forge; support osx-arm64; drop support for osx-64